+++ docs/example/redux/src/page/index/server/index.js
@@ -1,2 +1,3 @@
+import 'fast-text-encoding';
 import render from './render';
 global.render = render;